Fixes issues with running on LÖVE 0.8.x
On editor.lua, it tried saving files to a folder that didn't exist if you opened the SMB mappack. On enemy.lua, it had case-sensitivity issues. On several files, added the now-obsolete 0.8 :rewind() call. Changed error handler to redirect to the repository and made it actually display errors on 0.8.x.