Added the missing gray hardblock, fixed red pipe colors
The gray hardblock was present in SMB1, but is missing in Mari0, with a regular gray ground tile taking it's place. The colors on the mushroom, red pipe and red flagpole are consistent in SMB1, but the colors are inconsistent across all the "red" tiles in Mari0.

The gray hardblock replaces one of the 2 empty tiles at the bottom of the tileset, so it shouldn't break anything.